+++ main.js
@@ -10,22 +10,11 @@ export let users = [];
 
 import { fetchGet, fetchPost } from "./api.js";
 import renderComments from "./renderComments.js";
+import { currentDate } from "./data.js";
 
 const fetchPromise = () => {
   
   fetchGet()
-  
-  .then((response) => {
-
-    if (response.status === 500) {
-      
-      throw new Error('Сервер упал');
-
-    }
-
-   return response.json();
-
-  })
 
   .then((responseData) => {
 
@@ -70,32 +59,6 @@ const fetchPromise = () => {
 
 fetchPromise();
 
-
-
-export function currentDate(relevantDate) {
-
-const date = relevantDate;
-let day = date.getDate();
-let month = date.getMonth() + 1;
-let year = date.getFullYear() % 100;
-let hours = date.getHours();
-let minutes = date.getMinutes();
-
-if (day < 10) day = "0" + day;
-
-if (month < 10) month = "0" + month;
-
-if (year < 10) year = "0" + year;
-
-if (hours < 10) hours = "0" + hours;
-
-if (minutes < 10) minutes = "0" + minutes;
-
-let newDate = day + "." + month + "." + year + " " + hours + ":" + minutes;
-
-return newDate;
-
-}
 
 const initButtonsLikeListeners = () => {
 
@@ -182,24 +145,7 @@ buttonElement.addEventListener('click', () => {
   }
 
   fetchPost()
-  
-  .then((response) => {
 
-    if (response.status === 400) {
-
-    throw new Error('Ошибка ввода');
-
-    };
-
-    if (response.status === 500) {
-
-    throw new Error('Сервер упал');
-
-    };
-
-    return response.json();
-
-  })
   .then (() => {
       
     fetchPromise();
